requirements.txt
email-validator==2.0.0.post2
Jinja2==3.1.2
PyYAML==6.0.1
